+++ src/stores/useAuthStore.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
-import type { User } from '@/types/auth.types';
+import type { User } from '@/types/auth';
 
 interface AuthState {
   user: User | null;
