--- src/types/audio.ts
@@ -1,7 +1,0 @@
-export interface KitEvaluationRequest {
-  kitStageId: number;
-  evaluationScore: number;
-  evaluationFeedback: string;
-  isSuccess: boolean;
-  fileKey: string;
-}
