+++ src/types/auth/models.ts
@@ -1,0 +1,13 @@
+// ===== User & Auth Tokens =====
+export interface User {
+  id: number;
+  email: string;
+}
+
+export interface AuthTokens {
+  accessToken: string;
+  refreshToken: string;
+  tokenType: string;
+  expiresIn: number;
+  refreshExpiresIn: number;
+}
