+++ src/types/common/api.types.ts
@@ -1,0 +1,10 @@
+// ===== Error Response =====
+export interface ErrorResponse {
+  success: false;
+  status: number;
+  message: string;
+  timestamp: string;
+  error: {
+    code: string;
+  };
+}
