+++ src/types/lipsound/index.ts
@@ -1,0 +1,2 @@
+export * from './models';
+export * from './api.types';
