+++ src/types/situation/api.types.ts
@@ -1,16 +1,4 @@
-// 상황극 단일 항목
-export interface Situation {
-  situationId: number;
-  situationName: string;
-}
-
-// 상황극 상세 정보
-export interface SituationDetail {
-  situationId: number;
-  situationName: string;
-  description: string;
-  image: string;
-}
+import type { Situation, SituationDetail } from './models';
 
 // 상황극 목록 조회 응답
 export interface SituationListResponse {
