+++ src/types/talkingkit/models.ts
@@ -1,0 +1,8 @@
+export interface TalkingKit {
+  id: number;
+  category: string;
+  highlightedText: string;
+  mainText: string;
+  kitLabel: string;
+  isBookmarked: boolean;
+}
